前台基本增删改查完成.任务调度管理完成.剩余各短信发送接口的处理.cmppSocket连接的管理等

+++ src/main/java/com/zuma/sms/service/NumberGroupTypeService.java
@@ -1,5 +1,6 @@
 package com.zuma.sms.service;
 
+import com.zuma.sms.entity.NumberGroup;
 import com.zuma.sms.factory.PageRequestFactory;
 import com.zuma.sms.form.NumberGroupTypeForm;
 import com.zuma.sms.config.store.ConfigStore;
@@ -8,12 +9,16 @@ import com.zuma.sms.dto.PageVO;
 import com.zuma.sms.entity.NumberGroupType;
 import com.zuma.sms.enums.system.ErrorEnum;
 import com.zuma.sms.exception.SmsSenderException;
+import com.zuma.sms.repository.NumberGroupRepository;
 import com.zuma.sms.repository.NumberGroupTypeRepository;
 import org.springframework.beans.BeanUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
+import java.util.List;
 
 /**
  * author:ZhengXing
@@ -24,6 +29,8 @@ import org.springframework.stereotype.Service;
 public class NumberGroupTypeService {
 	@Autowired
 	private NumberGroupTypeRepository numberGroupTypeRepository;
+	@Autowired
+	private NumberGroupRepository numberGroupRepository;
 	@Autowired
 	private ConfigStore configStore;
 	@Autowired
@@ -59,6 +66,9 @@ public class NumberGroupTypeService {
 	 * 批量删除
 	 */
 	public void batchDelete(Long[] ids){
+		Page<NumberGroup> page = numberGroupRepository.findByTypeIdIn(ids, pageRequestFactory.buildForLimitOne());
+		if (page.getTotalElements() > 0)
+			throw new SmsSenderException("类别下存在号码组,无法删除");
 		numberGroupTypeRepository.deleteAllByIdIn(ids);
 	}
 
@@ -75,11 +85,22 @@ public class NumberGroupTypeService {
 	/**
 	 * 修改
 	 */
+	@Transactional
 	public void update(NumberGroupTypeForm form){
 		NumberGroupType numberGroupType = numberGroupTypeRepository.findOne(form.getId());
 		BeanUtils.copyProperties(form,numberGroupType);
 
+		//修改关联号码组的号码组类别名信息
+		numberGroupRepository.updateTypeNameByTypeId(form.getName(),form.getId());
+
 		numberGroupTypeRepository.save(numberGroupType);
+	}
+
+	/**
+	 * 查询所有
+	 */
+	public List<NumberGroupType> listAll() {
+		return numberGroupTypeRepository.findAll(pageRequestFactory.buildSort());
 	}
